backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pytesseract==0.3.13
Pillow==10.4.0
requests==2.32.3
gunicorn==22.0.0
